accepting symbols and removing syms/relocs

+++ elf-work/include/getsymstrtbl.h
@@ -1,6 +1,6 @@
 #ifndef GETSYMSTRTBL_H_
 #define GETSYMSTRTBL_H_
 
-char *get_symstrtbl(void *, Elf64_Shdr *, char *, int);
+char *get_symstrtbl(void *, Elf64_Shdr *, char *, short);
 
 #endif
